requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.3.0
mongoengine==0.27.0
PyJWT==2.3.0
pykerberos==1.2.4
pymongo==3.12.3
python-snappy==0.6.1
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
